backend/requirements.txt
alembic==1.5.6
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1


PyJWT~=2.0.1
DateTime~=4.3
pip~=21.0.1
argparse~=1.4.0
colorama~=0.4.4
setuptools~=49.2.1